+++ test/layout.test.ts
@@ -12,6 +12,27 @@ Deno.test("Layout tag", async () => {
   });
 });
 
+Deno.test("Layout tag (with filters)", async () => {
+  await test({
+    template: `
+    {{ layout "/my-file.tmpl" |> toUpperCase }}Hello world{{ /layout }}
+    `,
+    expected: "<h1>HELLO WORLD</h1>",
+    includes: {
+      "/my-file.tmpl": "<h1>{{ content }}</h1>",
+    },
+  });
+  await test({
+    template: `
+    {{ layout "/my-file.tmpl" |> toUpperCase }}Hello world{{ /layout }}
+    `,
+    expected: "<h1>hello world</h1>",
+    includes: {
+      "/my-file.tmpl": "<h1>{{ content |> toLowerCase }}</h1>",
+    },
+  });
+});
+
 Deno.test("Layout tag (with extra data)", async () => {
   await test({
     template: `
@@ -20,6 +41,20 @@ Deno.test("Layout tag (with extra data)", async () => {
     } }}Hello world{{ /layout }}
     `,
     expected: "<h1>Hello world</h1>",
+    includes: {
+      "/my-file.tmpl": "<{{ tag }}>{{ content }}</{{ tag }}>",
+    },
+  });
+});
+
+Deno.test("Layout tag (with extra data and filters)", async () => {
+  await test({
+    template: `
+    {{ layout "/my-file.tmpl" {
+      tag: "h1"
+    } |> toUpperCase }}Hello world{{ /layout }}
+    `,
+    expected: "<h1>HELLO WORLD</h1>",
     includes: {
       "/my-file.tmpl": "<{{ tag }}>{{ content }}</{{ tag }}>",
     },